FIX #7 verify access token function
